requirements.txt
langchain_core==0.3.66
langchain_huggingface==0.3.0
python-dotenv==1.1.0
torch==2.5.1+cu118
transformers==4.52.4
